Admin de usuarios done

+++ src/app/administrador/editar-usuario/editar-usuario.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Usuario, _Usuario } from 'src/models';
-import { SIDVIServices } from 'src/api';
+import { SIDVIServices, Defaults } from 'src/api';
 import Swal from 'sweetalert2';
 import { _APIResponse } from 'src/api/APIResponse';
+import { DomSanitizer } from '@angular/platform-browser';
 
 @Component({
   selector: 'app-editar-usuario',
@@ -12,17 +13,112 @@ import { _APIResponse } from 'src/api/APIResponse';
 export class EditarUsuarioComponent implements OnInit {
 
   usuarios: Usuario[];
+  usuarioModal: Usuario = new Usuario();
+  editar: boolean;
+  contrasena: string;
 
-  constructor(private sidvi: SIDVIServices) { }
+  constructor(private sidvi: SIDVIServices, private sanitizer: DomSanitizer) { }
 
   ngOnInit() {
     this.listarUsuarios();
   }
 
   listarUsuarios() {
-    this.sidvi.usuario.listarUsuarios(undefined, undefined, undefined, undefined, _Usuario.Rol.USUARIO).subscribe( usuarios => {
+    this.sidvi.usuario.listarUsuarios(undefined, undefined, undefined, undefined, _Usuario.Rol.USUARIO).subscribe(usuarios => {
       this.usuarios = usuarios.resultados.map((item: any) => new Usuario(item));
+
+      for (const administrador of this.usuarios) {
+        // Obtener imagen
+        if (Defaults.allowBase64Types.includes(this.usuarioModal.mimetypeFoto)) {
+          this.usuarioModal.archivoIconoImg = this.sanitizer.bypassSecurityTrustResourceUrl(this.usuarioModal.archivoFoto as string);
+        }
+      }
     });
+  }
+
+  detallesModal(usuario?: Usuario) {
+    if (usuario != null) {
+      this.usuarioModal = new Usuario(usuario);
+      this.editar = true;
+    } else {
+      this.usuarioModal = new Usuario();
+      this.usuarioModal.rol = _Usuario.Rol.USUARIO;
+      this.editar = false;
+      this.contrasena = '';
+    }
+  }
+
+  handleFileInput(files: FileList) {
+    if (files[0] != null) {
+      this.usuarioModal.localFile = files;
+      this.usuarioModal.localFileName = files[0].name;
+    }
+  }
+
+  editarUsuario() {
+    console.log(this.usuarioModal);
+    delete this.usuarioModal.archivoFoto;
+    delete this.usuarioModal.mimetypeFoto;
+
+    this.sidvi.usuario.actualizarUsuario(this.usuarioModal.idUsuario, this.usuarioModal).subscribe(
+      res => {
+        // Checar si se subió un doc
+        if (this.usuarioModal.localFile != null) {
+          this.actualizarUsuarioImagen('actualizado');
+          return;
+        }
+
+        Swal.fire({ title: '¡Listo!', text: 'Usuario actualizado correctamente', icon: 'success', heightAuto: false }).then((result) => {
+          if (result.value) {
+            this.listarUsuarios();
+          }
+        });
+      },
+      error => console.error(error)
+    );
+  }
+
+  guardarUsuario() {
+    delete this.usuarioModal.archivoFoto;
+    delete this.usuarioModal.mimetypeFoto;
+
+    if (this.usuarioModal.contrasena !== this.contrasena) {
+      console.error('Las contraseñas no coinciden');
+    }
+
+    this.sidvi.usuario.crearUsuario(this.usuarioModal).subscribe(
+      res => {
+        this.usuarioModal.idUsuario = res.extra.insertedId;
+
+        // Checar si se subió un doc
+        if (this.usuarioModal.localFile != null) {
+          this.actualizarUsuarioImagen('creado');
+          return;
+        }
+
+        Swal.fire({ title: '¡Listo!', text: 'Usuario creado correctamente', icon: 'success', heightAuto: false }).then((result) => {
+          if (result.value) {
+            this.listarUsuarios();
+          }
+        });
+      },
+      error => console.error(error)
+    );
+  }
+
+  actualizarUsuarioImagen(verbo: string) {
+    this.sidvi.usuario.cargarUsuarioFoto(this.usuarioModal.idUsuario, this.usuarioModal.localFile[0]).subscribe(
+      res => {
+        Swal.fire({ title: '¡Listo!', text: 'Usuario ' + verbo + ' correctamente', icon: 'success', heightAuto: false }).then((result) => {
+          if (result.value) {
+            this.listarUsuarios();
+          }
+        });
+      },
+      error => {
+        alert('Los archivos no se pudieron actualizar');
+      }
+    );
   }
 
   eliminar(usuario: Usuario) {
@@ -37,27 +133,33 @@ export class EditarUsuarioComponent implements OnInit {
       confirmButtonText: 'Si, eliminar',
       cancelButtonText: 'Cancelar',
       heightAuto: false
-  }).then((result) => {
+    }).then((result) => {
       if (result.value) {
-          this.sidvi.usuario.eliminarUsuario(usuario.idUsuario).subscribe(
-              res => {
-                  if (res.type === _APIResponse.TypeEnum.SUCCESS) {
-                      Swal.fire({
-                          title: 'Borrado completo',
-                          text: 'El bloque de información ha sido eliminado exitosamente',
-                          icon: 'success',
-                          heightAuto: false
-                      }).then((results) => {
-                          this.listarUsuarios();
-                      });
-                  }
-              },
-              error => {
-                  alert('No se pudo eliminar');
-              }
-          );
+
+        if (this.usuarios.length === 1) {
+          Swal.fire('Error', 'Debe haber mínimo un usuario en el sistema', 'error');
+          return;
+        }
+
+        this.sidvi.usuario.eliminarUsuario(usuario.idUsuario).subscribe(
+          res => {
+            if (res.type === _APIResponse.TypeEnum.SUCCESS) {
+              Swal.fire({
+                title: 'Borrado completo',
+                text: 'El bloque de información ha sido eliminado exitosamente',
+                icon: 'success',
+                heightAuto: false
+              }).then((results) => {
+                this.listarUsuarios();
+              });
+            }
+          },
+          error => {
+            alert('No se pudo eliminar');
+          }
+        );
       }
-  });
+    });
   }
 
 }
